- add client java teamobi

+++ src/main/java/nro/data/DataGame.java
@@ -124,7 +124,7 @@ public class DataGame {
         try {
             msg = Service.getInstance().messageNotMap((byte) 6);
             msg.writer().writeByte(vsMap);
-            msg.writer().writeByte(Manager.MAP_TEMPLATES.length);
+            msg.writer().writeShort(Manager.MAP_TEMPLATES.length);
             for (MapTemplate temp : Manager.MAP_TEMPLATES) {
                 msg.writer().writeUTF(temp.name);
             }
@@ -136,11 +136,11 @@ public class DataGame {
                 msg.writer().writeShort(temp.leg);
                 msg.writer().writeByte(0);
             }
-            msg.writer().writeByte(Manager.MOB_TEMPLATES.size());
+            msg.writer().writeShort(Manager.MOB_TEMPLATES.size());
             for (MobTemplate temp : Manager.MOB_TEMPLATES) {
                 msg.writer().writeByte(temp.type);
                 msg.writer().writeUTF(temp.name);
-                msg.writer().writeInt(temp.hp);
+                msg.writer().writeLong(temp.hp);
                 msg.writer().writeByte(temp.rangeMove);
                 msg.writer().writeByte(temp.speed);
                 msg.writer().writeByte(temp.dartType);
