+++ src/main/java/nro/models/mob/MobTemplate.java
@@ -13,7 +13,7 @@ public class MobTemplate {
     public int id;
     public byte type;
     public String name;
-    public int hp;
+    public long hp;
     public byte rangeMove;
     public byte speed;
     public byte dartType;
